Added newlines, added check if pid exists

+++ hu.bme.mit.incqueryd.runtime/scripts/stop-coordinator.sh
@@ -2,4 +2,4 @@
 
 cd "$( cd "$( dirname "$0" )" && pwd )"
 
-kill -9 $(<coordinator.pid)
+[ ! -f coordinator.pid ] && kill -9 $(<coordinator.pid)
